+++ src/components/inputs/FormField/FormField.tsx
@@ -10,8 +10,8 @@ import "./field.scss";
 
 export interface FormFieldInterface {
   name: string;
-  errors?: FormikErrors<{ email: string }>;
-  touched?: FormikTouched<{ email: string }>;
+  errors?: FormikErrors<any>;
+  touched?: FormikTouched<any>;
   readonly?: boolean;
   className?: string;
   noBorders?: boolean;